@@ -17,8 +17,8 @@
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = "1.4.32"
-    ext.beagle_version = "1.7.0"
+    ext.kotlin_version = "1.5.21"
+    ext.beagle_version = "2.0.0-rc1"
     repositories {
         google()
         jcenter()
