@@ -17,11 +17,8 @@
 
 package br.com.zup.beagle.scaffold
 
-import android.app.Application
-import br.com.zup.beagle.analytics.Analytics
 import br.com.zup.beagle.android.action.Action
-import br.com.zup.beagle.android.action.FormLocalActionHandler
-import br.com.zup.beagle.android.components.form.core.ValidatorHandler
+import br.com.zup.beagle.android.analytics.AnalyticsProvider
 import br.com.zup.beagle.android.data.serializer.adapter.generic.TypeAdapterResolver
 import br.com.zup.beagle.android.imagedownloader.BeagleImageDownloader
 import br.com.zup.beagle.android.logger.BeagleLogger
@@ -29,47 +26,36 @@ import br.com.zup.beagle.android.navigation.BeagleControllerReference
 import br.com.zup.beagle.android.navigation.DeepLinkHandler
 import br.com.zup.beagle.android.networking.HttpClient
 import br.com.zup.beagle.android.networking.HttpClientFactory
+import br.com.zup.beagle.android.networking.ViewClient
 import br.com.zup.beagle.android.networking.urlbuilder.UrlBuilder
 import br.com.zup.beagle.android.operation.Operation
 import br.com.zup.beagle.android.setup.BeagleConfig
 import br.com.zup.beagle.android.setup.BeagleSdk
 import br.com.zup.beagle.android.setup.DesignSystem
-import br.com.zup.beagle.android.store.StoreHandler
-import br.com.zup.beagle.android.view.BeagleActivity
 import br.com.zup.beagle.android.widget.WidgetView
-import br.com.zup.beagle.defaults.cache.StoreHandlerDefault
 import br.com.zup.beagle.defaults.httpclient.HttpClientDefault
 import br.com.zup.beagle.defaults.logger.BeagleLoggerDefault
-import br.com.zup.beagle.newanalytics.AnalyticsProvider
 
 class BeagleScaffold(private val beagleSdk: BeagleSdk) : BeagleSdk {
-    override val analytics: Analytics? = beagleSdk.analytics
-    override val analyticsProvider: AnalyticsProvider? = beagleSdk.analyticsProvider
+
     override val config: BeagleConfig = beagleSdk.config
-    override val controllerReference: BeagleControllerReference? = beagleSdk.controllerReference
     override val deepLinkHandler: DeepLinkHandler? = beagleSdk.deepLinkHandler
+    override val httpClientFactory: HttpClientFactory = beagleSdk.httpClientFactory ?: object :
+        HttpClientFactory {
+        override fun create(): HttpClient = HttpClientDefault()
+    }
     override val designSystem: DesignSystem? = beagleSdk.designSystem
-    override val formLocalActionHandler: FormLocalActionHandler? = beagleSdk.formLocalActionHandler
-    override val httpClient: HttpClient = beagleSdk.httpClient ?: HttpClientDefault()
-    override val httpClientFactory: HttpClientFactory? = beagleSdk.httpClientFactory
     override val imageDownloader: BeagleImageDownloader? = beagleSdk.imageDownloader
-    override val logger: BeagleLogger = beagleSdk.logger ?: BeagleLoggerDefault()
-    override val serverDrivenActivity: Class<BeagleActivity> = beagleSdk.serverDrivenActivity
-    override var storeHandler: StoreHandler? = beagleSdk.storeHandler
+    override val viewClient: ViewClient? = beagleSdk.viewClient
+    override val controllerReference: BeagleControllerReference? = beagleSdk.controllerReference
     override val typeAdapterResolver: TypeAdapterResolver? = beagleSdk.typeAdapterResolver
+    override val analyticsProvider: AnalyticsProvider? = beagleSdk.analyticsProvider
     override val urlBuilder: UrlBuilder? = beagleSdk.urlBuilder
-    override val validatorHandler: ValidatorHandler? = beagleSdk.validatorHandler
+    override val logger: BeagleLogger = beagleSdk.logger ?: BeagleLoggerDefault()
 
     override fun registeredActions(): List<Class<Action>> = beagleSdk.registeredActions()
 
     override fun registeredOperations(): Map<String, Operation> = beagleSdk.registeredOperations()
 
     override fun registeredWidgets(): List<Class<WidgetView>> = beagleSdk.registeredWidgets()
-
-    override fun init(application: Application) {
-        if (storeHandler == null) {
-            storeHandler = StoreHandlerDefault.newInstance(application)
-        }
-        super.init(application)
-    }
 }
