@@ -20,14 +20,13 @@ package br.com.zup.beagle.scaffold
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import br.com.zup.beagle.android.networking.RequestData
 import br.com.zup.beagle.android.utils.newServerDrivenIntent
 import br.com.zup.beagle.android.view.BeagleActivity
-import br.com.zup.beagle.android.view.ScreenRequest
 import br.com.zup.beagle.android.view.ServerDrivenActivity
 import br.com.zup.beagle.test.rules.BeagleComponentsRule
 import io.mockk.*
 import org.junit.Rule
-import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
@@ -46,19 +45,19 @@ internal class BeagleIntentTest {
         fun testToSampleReturnsAnIntent() {
             //Given
             val url = "http://10.0.2.2:8080/navigate-actions"
-            val screenRequest = ScreenRequest(url)
+            val requestData = RequestData(url)
             mockkConstructor(Intent::class)
             mockkStatic("br.com.zup.beagle.android.utils.ServerDrivenFactoryKt")
             mockkObject(BeagleActivity)
 
             every { anyConstructed<Intent>().putExtras(any<Bundle>()) } returns mockk()
-            every { BeagleActivity.bundleOf(screenRequest) } returns mockk()
+            every { BeagleActivity.bundleOf(requestData) } returns mockk()
 
             val activity = mockk<AppCompatActivity>(relaxUnitFun = true)
-            val beagleIntent = BeagleIntent(activity, screenRequest)
+            val beagleIntent = BeagleIntent(activity, requestData)
             every {
                 activity.newServerDrivenIntent<ServerDrivenActivity>(
-                    screenRequest
+                    requestData
                 )
             } returns mockk()
 
@@ -68,7 +67,7 @@ internal class BeagleIntentTest {
             //Then
             verify(exactly = 1) {
                 activity.newServerDrivenIntent<ServerDrivenActivity>(
-                    screenRequest
+                    requestData
                 )
             }
         }
