@@ -17,13 +17,9 @@
 
 package br.com.zup.beagle.scaffold
 
-import android.app.Application
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.runner.AndroidJUnit4
-import br.com.zup.beagle.analytics.Analytics
 import br.com.zup.beagle.android.action.Action
-import br.com.zup.beagle.android.action.FormLocalActionHandler
-import br.com.zup.beagle.android.components.form.core.ValidatorHandler
+import br.com.zup.beagle.android.analytics.AnalyticsProvider
 import br.com.zup.beagle.android.data.serializer.adapter.generic.TypeAdapterResolver
 import br.com.zup.beagle.android.imagedownloader.BeagleImageDownloader
 import br.com.zup.beagle.android.logger.BeagleLogger
@@ -31,24 +27,19 @@ import br.com.zup.beagle.android.navigation.BeagleControllerReference
 import br.com.zup.beagle.android.navigation.DeepLinkHandler
 import br.com.zup.beagle.android.networking.HttpClient
 import br.com.zup.beagle.android.networking.HttpClientFactory
+import br.com.zup.beagle.android.networking.ViewClient
 import br.com.zup.beagle.android.networking.urlbuilder.UrlBuilder
 import br.com.zup.beagle.android.operation.Operation
 import br.com.zup.beagle.android.setup.BeagleConfig
 import br.com.zup.beagle.android.setup.BeagleSdk
 import br.com.zup.beagle.android.setup.DesignSystem
-import br.com.zup.beagle.android.store.StoreHandler
-import br.com.zup.beagle.android.view.BeagleActivity
-import br.com.zup.beagle.android.view.ServerDrivenActivity
 import br.com.zup.beagle.android.widget.WidgetView
-import br.com.zup.beagle.defaults.cache.StoreHandlerDefault
 import br.com.zup.beagle.defaults.httpclient.HttpClientDefault
 import br.com.zup.beagle.defaults.logger.BeagleLoggerDefault
-import br.com.zup.beagle.newanalytics.AnalyticsProvider
 import br.com.zup.beagle.test.rules.BeagleComponentsRule
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.mockkObject
-import io.mockk.verify
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.jupiter.api.Assertions
@@ -60,54 +51,36 @@ internal class BeagleScaffoldTest {
 
     @get:Rule
     val begleComponentsRule = BeagleComponentsRule()
-    val application = ApplicationProvider.getApplicationContext() as Application
-    val analytics = mockk<Analytics>()
     val config = mockk<BeagleConfig>()
     val controllerReference = mockk<BeagleControllerReference>()
     val deepLinkHandler = mockk<DeepLinkHandler>()
     val designSystem = mockk<DesignSystem>()
-    val formLocalActionHandler = mockk<FormLocalActionHandler>()
     val httpClient = mockk<HttpClient>()
+    val viewClient = mockk<ViewClient>()
     val imageDownloader = mockk<BeagleImageDownloader>()
     val logger = mockk<BeagleLogger>()
-    val serverDrivenActivity: Class<BeagleActivity> =
-        ServerDrivenActivity::class.java as Class<BeagleActivity>
-    val storeHandler = mockk<StoreHandler>()
     val typeAdapterResolver = mockk<TypeAdapterResolver>(relaxed = true)
     val urlBuilder = mockk<UrlBuilder>()
-    val validatorHandler = mockk<ValidatorHandler>()
     val analyticsProvider = mockk<AnalyticsProvider>()
     val httpClientFactory = mockk<HttpClientFactory>()
+
+    @Before
+    fun setup() {
+        every { httpClientFactory.create() } returns httpClient
+    }
 
     @DisplayName("Then all its attributes are set equaling the FakeSdk ones")
     @Test
     fun beagleScaffoldSdkIsCreatedWithAllAttributesNotNull() {
         //Given
 
-        val beagleSdkFake = BeagleSdkFake(
-            analytics,
-            config,
-            controllerReference,
-            deepLinkHandler,
-            designSystem,
-            formLocalActionHandler,
-            httpClient,
-            imageDownloader,
-            logger,
-            serverDrivenActivity,
-            storeHandler,
-            typeAdapterResolver,
-            urlBuilder,
-            validatorHandler,
-            analyticsProvider,
-            httpClientFactory
-        )
+        val beagleSdkFake = getSdkFake(httpClientFactory = httpClientFactory, logger = logger)
         //When
         val beagleScaffold = BeagleScaffold(beagleSdkFake)
 
         //Then
 
-        Assertions.assertEquals(beagleSdkFake.analytics, beagleScaffold.analytics)
+        Assertions.assertEquals(beagleSdkFake.analyticsProvider, beagleScaffold.analyticsProvider)
         Assertions.assertEquals(beagleSdkFake.config, beagleScaffold.config)
         Assertions.assertEquals(
             beagleSdkFake.controllerReference,
@@ -115,24 +88,15 @@ internal class BeagleScaffoldTest {
         )
         Assertions.assertEquals(beagleSdkFake.deepLinkHandler, beagleScaffold.deepLinkHandler)
         Assertions.assertEquals(beagleSdkFake.designSystem, beagleScaffold.designSystem)
-        Assertions.assertEquals(
-            beagleSdkFake.formLocalActionHandler,
-            beagleScaffold.formLocalActionHandler
-        )
-        Assertions.assertEquals(beagleSdkFake.httpClient, beagleScaffold.httpClient)
+
         Assertions.assertEquals(beagleSdkFake.imageDownloader, beagleScaffold.imageDownloader)
         Assertions.assertEquals(beagleSdkFake.logger, beagleScaffold.logger)
-        Assertions.assertEquals(
-            beagleSdkFake.serverDrivenActivity,
-            beagleScaffold.serverDrivenActivity
-        )
-        Assertions.assertEquals(beagleSdkFake.storeHandler, beagleScaffold.storeHandler)
+
         Assertions.assertEquals(
             beagleSdkFake.typeAdapterResolver,
             beagleScaffold.typeAdapterResolver
         )
         Assertions.assertEquals(beagleSdkFake.urlBuilder, beagleScaffold.urlBuilder)
-        Assertions.assertEquals(beagleSdkFake.validatorHandler, beagleScaffold.validatorHandler)
         Assertions.assertEquals(beagleSdkFake.analyticsProvider, beagleScaffold.analyticsProvider)
         Assertions.assertEquals(beagleSdkFake.httpClientFactory, beagleScaffold.httpClientFactory)
     }
@@ -141,7 +105,7 @@ internal class BeagleScaffoldTest {
     @Test
     fun beagleScaffoldSdkIsCreatedWithAddingDefaultClasses() {
         //Given
-        val beagleSdkFake = getSdkFake(null, null, null)
+        val beagleSdkFake = getSdkFake(null, null)
         //When
         val beagleScaffold = BeagleScaffold(beagleSdkFake)
 
@@ -149,71 +113,44 @@ internal class BeagleScaffoldTest {
 
         Assertions.assertEquals(
             HttpClientDefault::class.java,
-            beagleScaffold.httpClient?.javaClass
+            beagleScaffold.httpClientFactory.create().javaClass
         )
         Assertions.assertEquals(
             BeagleLoggerDefault::class.java,
-            beagleScaffold.logger?.javaClass
+            beagleScaffold.logger.javaClass
         )
-        Assertions.assertEquals(null, beagleScaffold.storeHandler)
-    }
-
-    @DisplayName("Then the storeHandler should not be null")
-    @Test
-    fun beagleScaffoldCallsInitWithAnStoreHandlerNotNull() {
-        //Given
-        val beagleSdkFake = getSdkFake(httpClient, logger, null)
-        val beagleScaffold = BeagleScaffold(beagleSdkFake)
-        val storeHandlerDefault = mockk<StoreHandlerDefault>()
-        mockkObject(StoreHandlerDefault)
-        every { StoreHandlerDefault.newInstance(application) } returns storeHandlerDefault
-        beagleScaffold.init(application)
-        //Then
-        verify(exactly = 1) { StoreHandlerDefault.newInstance(application) }
-        Assertions.assertEquals(storeHandlerDefault, beagleScaffold.storeHandler)
     }
 
     private fun getSdkFake(
-        httpClient: HttpClient?,
+        httpClientFactory: HttpClientFactory?,
         logger: BeagleLogger?,
-        storeHandler: StoreHandler?
     ): BeagleSdk = BeagleSdkFake(
-        analytics,
-        config,
-        controllerReference,
-        deepLinkHandler,
-        designSystem,
-        formLocalActionHandler,
-        httpClient,
-        imageDownloader,
-        logger,
-        serverDrivenActivity,
-        storeHandler,
-        typeAdapterResolver,
-        urlBuilder,
-        validatorHandler,
-        analyticsProvider,
-        httpClientFactory
+        config = config,
+        deepLinkHandler = deepLinkHandler,
+        httpClientFactory = httpClientFactory,
+        designSystem = designSystem,
+        imageDownloader = imageDownloader,
+        viewClient = viewClient,
+        controllerReference = controllerReference,
+        typeAdapterResolver = typeAdapterResolver,
+        analyticsProvider = analyticsProvider,
+        urlBuilder = urlBuilder,
+        logger = logger
     )
 }
 
 class BeagleSdkFake(
-    override val analytics: Analytics?,
     override val config: BeagleConfig,
-    override val controllerReference: BeagleControllerReference?,
     override val deepLinkHandler: DeepLinkHandler?,
+    override val httpClientFactory: HttpClientFactory?,
     override val designSystem: DesignSystem?,
-    override val formLocalActionHandler: FormLocalActionHandler?,
-    override val httpClient: HttpClient?,
     override val imageDownloader: BeagleImageDownloader?,
-    override val logger: BeagleLogger?,
-    override val serverDrivenActivity: Class<BeagleActivity>,
-    override val storeHandler: StoreHandler?,
+    override val viewClient: ViewClient?,
+    override val controllerReference: BeagleControllerReference?,
     override val typeAdapterResolver: TypeAdapterResolver?,
-    override val urlBuilder: UrlBuilder?,
-    override val validatorHandler: ValidatorHandler?,
     override val analyticsProvider: AnalyticsProvider?,
-    override val httpClientFactory: HttpClientFactory?
+    override val urlBuilder: UrlBuilder?,
+    override val logger: BeagleLogger?
 ) : BeagleSdk {
     override fun registeredActions(): List<Class<Action>> = listOf()
 
